implemented class RemoteChargerList to abstract away logic having to do with fetching chargers
